Optionally allow users to bypass blocked page with password
Added a password option to the configuration. When present, a password
dialog will be displayed on the blocked page. If a user enters the
correct password they will have access to the frontend for the remainder
of their session. Once the session cookie is removed the user will no
longer be authenticated.